CORE-1320 lateral movement of existing Apollo Error-related feature
- moved Apollo Plugin + Error logging => 'src/server/apollo.errorHandling.ts'
- integration Test Case for Apollo Error logging
- added TooManyRequestsError
